pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: Apache Software License",
]
name = "nomad-external-eln-integrations"
description = "3rd Party Integration packages"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Amir Golparvar", email = "amir.golparvar@physik.hu-berlin.de" },
]
maintainers = [
    { name = "Amir Golparvar", email = "amir.golparvar@physik.hu-berlin.de" },
]
license = { file = "LICENSE" }
dependencies = ["nomad-lab>=1.3.0", "nomad-openbis>=1.0"]

[project.urls]
Repository = "https://github.com/FAIRmat-NFDI/nomad-external-eln-integrations"

[project.optional-dependencies]
dev = ["ruff", "pytest", "structlog"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4


[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # isort
    "I",
    # pylint 
    "PL",
]

ignore = [
    "F401",
    "F403", # 'from module import *' used; unable to detect undefined names
    "E501",
    "PLR2004",
    "PLR0912",
    "PLR0913", # Too many arguments in function definition (8 > 5)
    "PLR0915",
    "PLW2901",
    "I001", # [*] Import block is un-sorted or un-formatted
]

fixable = ["ALL"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# this is entirely optional, you can remove this if you wish to
[tool.ruff.format]
# use single quotes for strings.
quote-style = "single"

# indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[project.entry-points.'nomad.plugin']
elabftwparser = "nomad_external_eln_integrations.parsers:elabftw_parser_entry_point"
chemotionparser = "nomad_external_eln_integrations.parsers:chemotion_parser_entry_point"
labfolderschema = "nomad_external_eln_integrations.schema_packages.labfolder:labfolder_schema"
elabftwschema = "nomad_external_eln_integrations.schema_packages.elabftw:elabftw_schema"
openbisschema = "nomad_external_eln_integrations.schema_packages.openbis:openbis_schema"
elabftwexample = "nomad_external_eln_integrations.example_uploads:elabftwexample"
[tool.cruft]
# Avoid updating workflow files, this leads to permissions issues
skip = [".github/*"]
